remove user eager fetching from login-log, and move user controller
